+++ test-all.sh
@@ -3,4 +3,5 @@ node test-on-readme.js
 node test-of-persistence.js
 node test-of-trigger.js
 node test-of-renaming.js
+node test-of-rescheming.js
 node test-of-performance.js